Add a bit of _my first test_
Adapted directly from the Learn Go with Tests book since it reads really well, but with a bit of a Rust slant. I've included a taste of some stuff to come (e.g. the string type part) - not sure if it's worth leaving this out at this point? I did toy with the idea of going into some of the interesting stuff `assert_eq!` is doing, but definitely way to early for that I think (the types passed to it are different but the PartialEq trait is used and it determines the funciton to make the comparison at compile time - hence why it's a macro).